+++ ride_assignments.py
@@ -1,0 +1,6 @@
+import pandas as pd
+import numpy as np
+
+def assign_riders():
+    # TODO
+    return None
